Version avec une file d'attente des messages.

+++ src/timer.h
@@ -7,9 +7,9 @@ class Timer
 	void (*m_func)();
 
 public:
-	static constexpr uint32_t PRESCALER = 1024;
-	static constexpr double FREQUENCY = F_CPU / (PRESCALER * 2.0);
-	static constexpr double PERIOD_US = 1.0e6 * (PRESCALER * 2.0) / F_CPU;
+	static constexpr uint16_t PRESCALER = 1024;
+	static constexpr float FREQUENCY = F_CPU / (PRESCALER * 2.0);
+	static constexpr float PERIOD_US = 1.0e6 * (PRESCALER * 2.0) / F_CPU;
 
 	void Init(void (*func)());
